requirements.txt
fastapi
uvicorn
pydantic
pyyaml
langchain
langchain-openai
requests
